Move mock classes to separate files and remove the hangfire server
